# Integration FlashInfer fused allreduce RMSNorm