--- a/vllm/config.py
+++ b/vllm/config.py
@@ -3962,8 +3962,10 @@ class PassConfig:
     """Whether to enable sequence parallelism."""
     enable_async_tp: bool = False
     """Whether to enable async TP."""
-    enable_flashinfer_allreduce_fusion: bool = False
+    enable_fi_allreduce_fusion: bool = False
     """Whether to enable flashinfer allreduce fusion."""
+    fi_allreduce_fusion_max_token_num: int = 1024
+    """Max number of tokens to used in flashinfer allreduce fusion."""
 
     # TODO(luka) better pass enabling system.