--- a/vllm/config.py
+++ b/vllm/config.py
@@ -3962,6 +3962,8 @@ class PassConfig:
     """Whether to enable sequence parallelism."""
     enable_async_tp: bool = False
     """Whether to enable async TP."""
+    enable_flashinfer_allreduce_fusion: bool = False
+    """Whether to enable flashinfer allreduce fusion."""
 
     # TODO(luka) better pass enabling system.