--- a/vllm/envs.py
+++ b/vllm/envs.py
@@ -139,7 +139,6 @@ if TYPE_CHECKING:
     VLLM_ROCM_QUICK_REDUCE_CAST_BF16_TO_FP16: bool = True
     VLLM_ROCM_QUICK_REDUCE_MAX_SIZE_BYTES_MB: Optional[int] = None
     VLLM_NIXL_ABORT_REQUEST_TIMEOUT: int = 120
-    VLLM_USE_FLASHINFER_ALLREDUCE: bool = False
 
 
 def get_default_cache_root():