… tests/compile/test_fusion_all_reduce.py