## Purpose

Integrates FlashInfer fused allreduce RMSNorm using fusion passes.
Can be enabled in compilation config: `--compilation-config='{"pass_config": {"enable_flashinfer_allreduce_fusion": true}, "custom_ops": ["+rms_norm"], "level":3}'`

Baseline, no custom ops:
![image](https://github.com/user-attachments/assets/656547ef-8205-47cc-bee1-c26e9abe5fe8)
After:
![image](https://github.com/user-attachments/assets/5a1c3820-5357-4554-a6a6-6db10b4ba09b)

## Benchmarking End-to-End
Llama-3.1-70B-Instruct TP=4 on B200 GPUs

### Client: 
       DURATION_SECONDS=60; \
       vllm bench serve \
        --model meta-llama/Llama-3.1-70B-Instruct \
        --dataset-name sonnet  \
        --dataset-path benchmarks/sonnet.txt \
        --request-rate "$qps" \
        --num-prompts $((DURATION_SECONDS * qps))`
### Server.

Baseline:
`vllm serve meta-llama/Llama-3.1-70B-Instruct  --disable-log-requests --no-enable-prefix-caching -tp 4`
PR: `vllm serve meta-llama/Llama-3.1-70B-Instruct  --disable-log-requests --no-enable-prefix-caching -tp 4 --compilation-config='{"pass_config": {"enable_flashinfer_allreduce_fusion": true}, "custom_ops": ["+rms_norm"], "level":3}'`

### Results:
Baseline
| QPS | Mean TTFT (ms) | Median TTFT (ms) | Mean TPOT (ms) | Median TPOT (ms) |
|---|---|---|---|---|
| 1 | 42.557 | 41.910 | 13.210 | 13.216 |
| 5 | 46.372 | 44.402 | 14.442 | 14.382 |
| 10 | 54.465 | 49.271 | 17.834 | 17.953 |  

PR:
| QPS | Mean TTFT (ms) | Median TTFT (ms) | Mean TPOT (ms) | Median TPOT (ms) |
|---|---|---|---|---|
| 1 | 43.9 | 42.54 | 11.875 | 11.864 |
| 5 | 48.234 | 46.041 | 13.261 | 13.185 |
| 10 | 52.725 | 46.457 | 15.352 | 14.292 |

TPOT gets around 10-15% speedup 

## Test Plan
Added `tests/compile/test_fusion_all_reduce.py`